fix: stubbing of run mode readdir
